Add release.nix for cross-builds on Hydra.
We now also accept the crossSystem in default.nix, so we can cross-build
for other architectures. Right now, we only build for Linux and Windows
system, or at least we *try* to. The output of the default.nix now also
uses crossDrv, so we don't get an unexpected native build here.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>